refactor: address additional review comments
- Change max_module_size from usize to u64 for consistency with other VMLimits fields
- Add safety documentation to from_precompiled explaining why no size check is needed
- Add boundary condition test for modules exactly at the size limit

+++ crates/runtime/src/logic.rs
@@ -110,7 +110,7 @@ const DEFAULT_MAX_BLOB_CHUNK_SIZE_MIB: u64 = 10;
 /// Default maximum method name length in bytes.
 const DEFAULT_MAX_METHOD_NAME_LENGTH: u64 = 256;
 /// Default maximum WASM module size in MiB (10 MiB).
-const DEFAULT_MAX_MODULE_SIZE_MIB: usize = 10;
+const DEFAULT_MAX_MODULE_SIZE_MIB: u64 = 10;
 
 /// Defines the resource limits for a VM instance.
 ///
@@ -121,7 +121,7 @@ pub struct VMLimits {
     /// The maximum size of a WASM module in bytes before compilation.
     /// This limit prevents memory exhaustion attacks from large malicious modules.
     /// Setting this to 0 will reject all modules.
-    pub max_module_size: usize,
+    pub max_module_size: u64,
     /// The maximum number of memory pages allowed.
     pub max_memory_pages: u32,
     /// The maximum stack size in bytes.
@@ -170,7 +170,7 @@ impl Default for VMLimits {
         }
 
         Self {
-            max_module_size: DEFAULT_MAX_MODULE_SIZE_MIB * ONE_MIB as usize,
+            max_module_size: DEFAULT_MAX_MODULE_SIZE_MIB as u64 * u64::from(ONE_MIB),
             max_memory_pages: ONE_KIB,
             max_stack_size: DEFAULT_MAX_STACK_SIZE_KIB * ONE_KIB as usize,
             max_registers: DEFAULT_MAX_REGISTERS,
